Pruebas hasta el momento

+++ src/main/groovy/forofiuba/Constantes.groovy
@@ -5,7 +5,7 @@ package forofiuba
 class Constantes {
     final static Integer PUNTAJE_INICIAL = 1
     final static Integer MISMA_CARRERA = 2
-    final static Integer CURSO_OTRA = 1.5
+    final static Float CURSO_OTRA = 1.5
     final static Integer PISO_RECOMENDABLE = 15
     final static Integer PISO_PROBABLE = 8
     final static String RECOMENDABLE = "La vas a cursar aca "
